fix: close() call on already closed S3WritableByteChannel (awslabs#453) (#454)

+++ src/main/java/software/amazon/nio/spi/s3/S3WritableByteChannel.java
@@ -87,6 +87,10 @@ class S3WritableByteChannel implements WritableByteChannel {
     @Override
     public void close() throws IOException {
         channel.close();
+        if (!open) {
+            // channel has already been closed -> close() should have no effect
+            return;
+        }
 
         s3TransferUtil.uploadLocalFile(path, tempFile);
         Files.deleteIfExists(tempFile);
